make Reader a final class
